ProductController przechowuje id grup i zamowien, nie encje.

+++ src/main/java/com/kodilla/ecommercee/domain/OrderEntity.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Table(name = "ORDERS")
 public final class OrderEntity {
 
-    private int orderId;
+    private Long orderId;
     private Date dateOfOrder = new Date();
     private List<ProductEntity> products = new ArrayList<>();
     private UserEntity userEntity;
@@ -21,7 +21,7 @@ public final class OrderEntity {
     @Id
     @GeneratedValue
     @Column(name = "ORDER_ID", unique = true, nullable = false)
-    public int getOrderId() {
+    public Long getOrderId() {
         return orderId;
     }
 
@@ -51,7 +51,7 @@ public final class OrderEntity {
         this.userEntity = userEntity;
     }
 
-    private void setOrderId(int orderId) {
+    private void setOrderId(Long orderId) {
         this.orderId = orderId;
     }
